Renamed BigWig.py to BigWigDatasource
