+++ oncotator/datasources/BigWigDatasource.py
